+++ src/features/data-upload/Uploader.module.css
@@ -1,0 +1,10 @@
+.wrapper{
+  display: flex;
+  flex-direction: column;
+}
+
+.upload{
+  display:block;
+  align-self: center;
+  margin-top: 50px;
+}
